refactor: 공지사항 페이지 useEffect 관련 호출 함수 분리
